+++ tests/Integration/Encoding/UtfTest.php
@@ -15,6 +15,8 @@ use oxDb;
 use oxEmail;
 use oxField;
 use oxGroups;
+use OxidEsales\Eshop\Core\Config;
+use OxidEsales\Eshop\Core\Registry;
 use OxidEsales\EshopCommunity\Application\Model\RssFeed;
 use oxLinks;
 use oxList;
@@ -1124,7 +1126,7 @@ class UtfTest extends \OxidTestCase
         $this->getConfig()->setConfigParam('bl_perfParseLongDescinSmarty', false);
 
         $rssFeed = oxNew('oxrssfeed');
-        $rssFeed->setConfig($config);
+        Registry::set(Config::class, $config);
 
         $shortDescription = 'agentūrų Литовские für';
         $longDescription = new stdClass();
@@ -1163,7 +1165,7 @@ class UtfTest extends \OxidTestCase
         $oShop->oxshops__oxname = new oxField($sValue);
         $oCfg->expects($this->any())->method('getActiveShop')->will($this->returnValue($oShop));
 
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals($sValue . '/Test', $oRss->UNITprepareFeedName('Test'));
     }
 
@@ -1195,7 +1197,7 @@ class UtfTest extends \OxidTestCase
                 }
             )
         );
-        $rssFeedMock->setConfig($config);
+        Registry::set(Config::class, $config);
 
         $expectedSearchArticleTitle = 'Test Shop/SEARCH_FOR_PRODUCTS_CATEGORY_VENDOR_MANUFACTURERtssscat' . $value . 'man';
         $this->assertEquals($expectedSearchArticleTitle, $rssFeedMock->getSearchArticlesTitle('tsss', 'cat', $value, 'man'));