+++ tests/Unit/Application/Controller/Admin/ModuleListTest.php
@@ -6,6 +6,8 @@
 namespace OxidEsales\EshopCommunity\Tests\Unit\Application\Controller\Admin;
 
 use oxConfig;
+use OxidEsales\Eshop\Core\Config;
+use OxidEsales\Eshop\Core\Registry;
 use PHPUnit\Framework\MockObject\MockObject;
 
 class ModuleListTest extends \OxidTestCase
@@ -28,7 +30,7 @@ class ModuleListTest extends \OxidTestCase
         $config->expects($this->any())->method('getModulesDir')->will($this->returnValue(__DIR__.'/../../../testData/modules/'));
 
         $oView = oxNew('Module_List');
-        $oView->setConfig($config);
+        Registry::set(Config::class, $config);
         $this->assertEquals('module_list.tpl', $oView->render());
 
         $aViewData = $oView->getViewData();