+++ tests/Unit/Application/Model/RssfeedTest.php
@@ -6,6 +6,8 @@
 namespace OxidEsales\EshopCommunity\Tests\Unit\Application\Model;
 
 use \oxField;
+use OxidEsales\Eshop\Core\Config;
+use OxidEsales\Eshop\Core\Registry;
 use oxRssFeed;
 use \stdClass;
 use \oxList;
@@ -41,7 +43,7 @@ class RssfeedTest extends \OxidTestCase
         $oCfg->expects($this->any())->method('getShopUrl')->will($this->returnValue("http://homeurl/"));
         $oCfg->expects($this->any())->method('getImageUrl')->will($this->returnValue("http://homeurl/lalala/"));
 
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $oRss->p_loadBaseChannel();
         $edition = strtolower($this->getTestConfig()->getShopEdition());
         $expect = array(
@@ -161,7 +163,7 @@ class RssfeedTest extends \OxidTestCase
         $oCfg = $this->getConfig();
         $oCfg->setConfigParam('aCurrencies', array('EUR@1.00@.@.@EUR@1'));
         $oRss = oxNew('oxRssFeed');
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
 
         $oLongDesc = new stdClass();
         $oLongDesc->value = "artlogndesc";
@@ -254,7 +256,7 @@ class RssfeedTest extends \OxidTestCase
         $oCfg = $this->getConfig();
         $oCfg->setConfigParam('aCurrencies', array('EUR@1.00@.@.@EUR@1'));
         $oRss = oxNew('oxRssFeed');
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
 
         $oArt1 = $this->getMock(\OxidEsales\Eshop\Application\Model\Article::class, array("getLink", "getLongDesc"));
         $oArt1->expects($this->any())->method('getLink')->will($this->returnValue("artlink"));
@@ -305,7 +307,7 @@ class RssfeedTest extends \OxidTestCase
         $oCfg->expects($this->any())->method('getActShopCurrencyObject')->will($this->returnValue($oActCur));
 
         $oRss = oxNew('oxRssFeed');
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
 
         $oLongDesc = new oxField("artlogndesc");
         
@@ -365,7 +367,7 @@ class RssfeedTest extends \OxidTestCase
         $oCfg->setConfigParam('aCurrencies', array('CHF@1.00@,@.@CHF@1'));
 
         $oRss = oxNew('oxRssFeed');
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
 
         $oLongDesc = new stdClass();
         $oLongDesc->value = "artlogndesc";
@@ -418,7 +420,7 @@ class RssfeedTest extends \OxidTestCase
         $oCfg->expects($this->any())->method('getShopUrl')->will($this->returnValue('http://homeurl/?'));
 
         $oRss = oxNew('oxrssfeed');
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals('http://homeurl/?cl=rss&amp;fnc=topshop&amp;lang=1extra', $oRss->UNITprepareUrl('cl=rss&amp;fnc=topshop', 'asd'));
     }
 
@@ -437,7 +439,7 @@ class RssfeedTest extends \OxidTestCase
         $oRss = $this->getMock(\OxidEsales\Eshop\Application\Model\RssFeed::class, array('getConfig'));
         $oRss->expects($this->any())->method('getConfig')->will($this->returnValue($oCfg));
 
-        \OxidEsales\Eshop\Core\Registry::getSeoEncoder()->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
 
         $this->assertEquals('http://homeurl/?cl=rss&amp;fnc=topshop&amp;lang=1 - SEO - rss/asd/extra', $oRss->UNITprepareUrl('cl=rss&amp;fnc=topshop', 'asd'));
     }
@@ -451,7 +453,7 @@ class RssfeedTest extends \OxidTestCase
         $oShop->oxshops__oxname = new oxField('Test Shop');
         $oCfg->expects($this->any())->method('getActiveShop')->will($this->returnValue($oShop));
 
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals('Test Shop/Test', $oRss->UNITprepareFeedName('Test'));
     }
 
@@ -463,28 +465,28 @@ class RssfeedTest extends \OxidTestCase
 
         oxTestModules::publicize('oxrssfeed', '_getShopUrl');
         $oRss = oxNew('oxRssFeed');
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
 
         $this->assertEquals('http://homeurl/?', $oRss->p_getShopUrl());
 
         $oCfg = $this->getMock(\OxidEsales\Eshop\Core\Config::class, array('getShopUrl'));
         $oCfg->expects($this->any())->method('getShopUrl')->will($this->returnValue("http://homeurl/"));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals('http://homeurl/?', $oRss->p_getShopUrl());
 
         $oCfg = $this->getMock(\OxidEsales\Eshop\Core\Config::class, array('getShopUrl'));
         $oCfg->expects($this->any())->method('getShopUrl')->will($this->returnValue("http://homeurl/?sdf"));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals('http://homeurl/?sdf&amp;', $oRss->p_getShopUrl());
 
         $oCfg = $this->getMock(\OxidEsales\Eshop\Core\Config::class, array('getShopUrl'));
         $oCfg->expects($this->any())->method('getShopUrl')->will($this->returnValue("http://homeurl/?sdf&"));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals('http://homeurl/?sdf&', $oRss->p_getShopUrl());
 
         $oCfg = $this->getMock(\OxidEsales\Eshop\Core\Config::class, array('getShopUrl'));
         $oCfg->expects($this->any())->method('getShopUrl')->will($this->returnValue("http://homeurl/?sdf&amp;"));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals('http://homeurl/?sdf&amp;', $oRss->p_getShopUrl());
     }
 
@@ -557,7 +559,7 @@ class RssfeedTest extends \OxidTestCase
         $oShop = oxNew('oxShop');
         $oShop->oxshops__oxname = new oxField('Test Shop');
         $oCfg->expects($this->any())->method('getActiveShop')->will($this->returnValue($oShop));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
 
         $this->assertEquals('Test Shop/TOP_OF_THE_SHOPtr', $oRss->getTopInShopTitle());
     }
@@ -614,7 +616,7 @@ class RssfeedTest extends \OxidTestCase
         $oShop = oxNew('oxShop');
         $oShop->oxshops__oxname = new oxField('Test Shop');
         $oCfg->expects($this->any())->method('getActiveShop')->will($this->returnValue($oShop));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals('Test Shop/NEWEST_SHOP_PRODUCTStr', $oRss->getNewestArticlesTitle());
     }
 
@@ -686,7 +688,7 @@ class RssfeedTest extends \OxidTestCase
         $oShop = oxNew('oxShop');
         $oShop->oxshops__oxname = new oxField('Test Shop');
         $oCfg->expects($this->any())->method('getActiveShop')->will($this->returnValue($oShop));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
 
         $oCat = oxNew('oxcategory');
         $oCat->load($sCatId);
@@ -792,7 +794,7 @@ class RssfeedTest extends \OxidTestCase
         $oShop = oxNew('oxShop');
         $oShop->oxshops__oxname = new oxField('Test Shop');
         $oCfg->expects($this->any())->method('getActiveShop')->will($this->returnValue($oShop));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals('Test Shop/SEARCH_FOR_PRODUCTS_CATEGORY_VENDOR_MANUFACTURERtssscatvendman', $oRss->getSearchArticlesTitle('tsss', 'cat', 'vend', 'man'));
     }
 
@@ -842,9 +844,6 @@ class RssfeedTest extends \OxidTestCase
     public function testLoadSearchArticles()
     {
         oxTestModules::addFunction('oxrssfeed', '_getSearchParamsUrl', '{ return "klnk"; }');
-        $oConfig = $this->getConfig();
-        $oRss = oxNew('oxRssFeed');
-        $oRss->setConfig($oConfig);
 
         $this->getConfig()->setConfigParam('iRssItemsCount', 50);
         oxTestModules::addFunction('oxLang', 'getBaseLanguage', '{return 1;}');
@@ -933,7 +932,7 @@ class RssfeedTest extends \OxidTestCase
         $oShop = oxNew('oxShop');
         $oShop->oxshops__oxname = new oxField('Test Shop');
         $oCfg->expects($this->any())->method('getActiveShop')->will($this->returnValue($oShop));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $oArt = oxNew('oxArticle');
         $oArt->oxarticles__oxtitle = new oxField('tsss');
         $this->assertEquals('Test Shop/LISTMANIA_LIST_FORtsss', $oRss->getRecommListsTitle($oArt));
@@ -1083,7 +1082,7 @@ class RssfeedTest extends \OxidTestCase
         $oShop = oxNew('oxShop');
         $oShop->oxshops__oxname = new oxField('Test Shop');
         $oCfg->expects($this->any())->method('getActiveShop')->will($this->returnValue($oShop));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $oRecommList = oxNew('oxRecommList');
         $oRecommList->oxrecommlists__oxtitle = new oxField('tsss');
         $this->assertEquals('Test Shop/LISTMANIA_LIST_PRODUCTStsss', $oRss->getRecommListArticlesTitle($oRecommList));
@@ -1190,7 +1189,7 @@ class RssfeedTest extends \OxidTestCase
         $oShop = oxNew('oxShop');
         $oShop->oxshops__oxname = new oxField('Test Shop');
         $oCfg->expects($this->any())->method('getActiveShop')->will($this->returnValue($oShop));
-        $oRss->setConfig($oCfg);
+        Registry::set(Config::class, $oCfg);
         $this->assertEquals('Test Shop/BARGAINtr', $oRss->getBargainTitle());
     }